162 access management   phase 3

+++ KinaUnaProgenyApi/Services/MeasurementService.cs
@@ -1,27 +1,35 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json;
 using System.Threading.Tasks;
 using KinaUna.Data;
 using KinaUna.Data.Contexts;
 using KinaUna.Data.Extensions;
 using KinaUna.Data.Models;
+using KinaUna.Data.Models.AccessManagement;
+using KinaUna.Data.Models.CacheManagement;
+using KinaUnaProgenyApi.Services.AccessManagementService;
+using KinaUnaProgenyApi.Services.CacheServices;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Distributed;
-using Newtonsoft.Json;
 
 namespace KinaUnaProgenyApi.Services
 {
     public class MeasurementService : IMeasurementService
     {
         private readonly ProgenyDbContext _context;
+        private readonly IAccessManagementService _accessManagementService;
         private readonly IDistributedCache _cache;
+        private readonly IKinaUnaCacheService _kinaUnaCacheService;
         private readonly DistributedCacheEntryOptions _cacheOptions = new();
         private readonly DistributedCacheEntryOptions _cacheOptionsSliding = new();
 
-        public MeasurementService(ProgenyDbContext context, IDistributedCache cache)
+        public MeasurementService(ProgenyDbContext context, IDistributedCache cache, IAccessManagementService accessManagementService, IKinaUnaCacheService kinaUnaCacheService)
         {
             _context = context;
+            _accessManagementService = accessManagementService;
             _cache = cache;
+            _kinaUnaCacheService = kinaUnaCacheService;
             _cacheOptions.SetAbsoluteExpiration(new System.TimeSpan(0, 5, 0)); // Expire after 5 minutes.
             _cacheOptionsSliding.SetSlidingExpiration(new System.TimeSpan(7, 0, 0, 0)); // Expire after a week.
         }
@@ -31,15 +39,25 @@ namespace KinaUnaProgenyApi.Services
         /// First tries to get the Measurement from the cache, then from the database if it's not in the cache.
         /// </summary>
         /// <param name="id">The MeasurementId of the Measurement entity to get.</param>
+        /// <param name="currentUserInfo">The UserInfo object for the current user, to check permissions.</param>
         /// <returns>The Measurement object with the given MeasurementId. Null if the Measurement doesn't exist.</returns>
-        public async Task<Measurement> GetMeasurement(int id)
+        public async Task<Measurement> GetMeasurement(int id, UserInfo currentUserInfo)
         {
+            if (!await _accessManagementService.HasItemPermission(KinaUnaTypes.TimeLineType.Measurement, id, currentUserInfo, PermissionLevel.View))
+            {
+                return null;
+            }
+
             Measurement measurement = await GetMeasurementFromCache(id);
             if (measurement == null || measurement.MeasurementId == 0)
             {
                 measurement = await SetMeasurementInCache(id);
             }
-
+            if (measurement == null || measurement.MeasurementId == 0)
+            {
+                return null;
+            }
+            measurement.ItemPerMission = await _accessManagementService.GetItemPermissionForUser(KinaUnaTypes.TimeLineType.Measurement, measurement.MeasurementId, measurement.ProgenyId, 0, currentUserInfo);
             return measurement;
         }
 
@@ -57,7 +75,7 @@ namespace KinaUnaProgenyApi.Services
                 return null;
             }
 
-            Measurement measurement = JsonConvert.DeserializeObject<Measurement>(cachedMeasurement);
+            Measurement measurement = JsonSerializer.Deserialize<Measurement>(cachedMeasurement, JsonSerializerOptions.Web);
             return measurement;
         }
 
@@ -71,7 +89,7 @@ namespace KinaUnaProgenyApi.Services
             Measurement measurement = await _context.MeasurementsDb.AsNoTracking().SingleOrDefaultAsync(m => m.MeasurementId == id);
             if (measurement == null) return null;
 
-            await _cache.SetStringAsync(Constants.AppName + Constants.ApiVersion + "measurement" + id, JsonConvert.SerializeObject(measurement), _cacheOptionsSliding);
+            await _cache.SetStringAsync(Constants.AppName + Constants.ApiVersion + "measurement" + id, JsonSerializer.Serialize(measurement, JsonSerializerOptions.Web), _cacheOptionsSliding);
 
             _ = await SetMeasurementsListInCache(measurement.ProgenyId);
 
@@ -82,15 +100,26 @@ namespace KinaUnaProgenyApi.Services
         /// Adds a new Measurement to the database and the cache.
         /// </summary>
         /// <param name="measurement">The Measurement object to add.</param>
+        /// <param name="currentUserInfo"></param>
         /// <returns>The added Measurement object.</returns>
-        public async Task<Measurement> AddMeasurement(Measurement measurement)
+        public async Task<Measurement> AddMeasurement(Measurement measurement, UserInfo currentUserInfo)
         {
+            if (!await _accessManagementService.HasProgenyPermission(measurement.ProgenyId, currentUserInfo, PermissionLevel.Add))
+            {
+                return null;
+            }
+
             Measurement measurementToAdd = new();
             measurementToAdd.CopyPropertiesForAdd(measurement);
 
             _ = _context.MeasurementsDb.Add(measurementToAdd);
             _ = await _context.SaveChangesAsync();
+
+            await _accessManagementService.AddItemPermissions(KinaUnaTypes.TimeLineType.Measurement, measurementToAdd.MeasurementId, measurementToAdd.ProgenyId, 0, measurementToAdd.ItemPermissionsDtoList,
+                currentUserInfo);
             _ = await SetMeasurementInCache(measurementToAdd.MeasurementId);
+
+            await _kinaUnaCacheService.SetProgenyOrFamilyTimelineUpdatedCache(measurementToAdd.ProgenyId, 0, KinaUnaTypes.TimeLineType.Measurement);
 
             return measurementToAdd;
         }
@@ -99,9 +128,15 @@ namespace KinaUnaProgenyApi.Services
         /// Updates a Measurement in the database and the cache.
         /// </summary>
         /// <param name="measurement">The Measurement with the updated properties.</param>
+        /// <param name="currentUserInfo"></param>
         /// <returns>The updated Measurement.</returns>
-        public async Task<Measurement> UpdateMeasurement(Measurement measurement)
+        public async Task<Measurement> UpdateMeasurement(Measurement measurement, UserInfo currentUserInfo)
         {
+            if (!await _accessManagementService.HasItemPermission(KinaUnaTypes.TimeLineType.Measurement, measurement.MeasurementId, currentUserInfo, PermissionLevel.Edit))
+            {
+                return null;
+            }
+
             Measurement measurementToUpdate = await _context.MeasurementsDb.SingleOrDefaultAsync(m => m.MeasurementId == measurement.MeasurementId);
             if (measurementToUpdate == null) return null;
 
@@ -110,8 +145,11 @@ namespace KinaUnaProgenyApi.Services
             _ = _context.MeasurementsDb.Update(measurementToUpdate);
             _ = await _context.SaveChangesAsync();
 
+            await _accessManagementService.UpdateItemPermissions(KinaUnaTypes.TimeLineType.Measurement, measurementToUpdate.MeasurementId, measurementToUpdate.ProgenyId, 0, measurementToUpdate.ItemPermissionsDtoList,
+                currentUserInfo);
             _ = await SetMeasurementInCache(measurement.MeasurementId);
 
+            await _kinaUnaCacheService.SetProgenyOrFamilyTimelineUpdatedCache(measurementToUpdate.ProgenyId, 0, KinaUnaTypes.TimeLineType.Measurement);
             return measurementToUpdate;
         }
 
@@ -119,17 +157,31 @@ namespace KinaUnaProgenyApi.Services
         /// Deletes a Measurement from the database and the cache.
         /// </summary>
         /// <param name="measurement">The Measurement to delete.</param>
+        /// <param name="currentUserInfo"></param>
         /// <returns>The deleted Measurement object.</returns>
-        public async Task<Measurement> DeleteMeasurement(Measurement measurement)
+        public async Task<Measurement> DeleteMeasurement(Measurement measurement, UserInfo currentUserInfo)
         {
+            if (!await _accessManagementService.HasItemPermission(KinaUnaTypes.TimeLineType.Measurement, measurement.MeasurementId, currentUserInfo, PermissionLevel.Admin))
+            {
+                return null;
+            }
+
             Measurement measurementToDelete = await _context.MeasurementsDb.SingleOrDefaultAsync(m => m.MeasurementId == measurement.MeasurementId);
             if (measurementToDelete == null) return null;
 
             _ = _context.MeasurementsDb.Remove(measurementToDelete);
             _ = await _context.SaveChangesAsync();
 
+            // Remove all associated permissions.
+            List<TimelineItemPermission> timelineItemPermissionsList = await _accessManagementService.GetTimelineItemPermissionsList(KinaUnaTypes.TimeLineType.Contact, measurementToDelete.MeasurementId, currentUserInfo);
+            foreach (TimelineItemPermission permission in timelineItemPermissionsList)
+            {
+                await _accessManagementService.RevokeItemPermission(permission, currentUserInfo);
+            }
+
             await RemoveMeasurementFromCache(measurement.MeasurementId, measurement.ProgenyId);
 
+            await _kinaUnaCacheService.SetProgenyOrFamilyTimelineUpdatedCache(measurementToDelete.ProgenyId, 0, KinaUnaTypes.TimeLineType.Measurement);
 
             return measurement;
         }
@@ -152,19 +204,39 @@ namespace KinaUnaProgenyApi.Services
         /// First tries to get the list from the cache, then from the database if it's not in the cache.
         /// </summary>
         /// <param name="progenyId">The ProgenyId of the Progeny to get Measurements for.</param>
-        /// <param name="accessLevel">The access level of the user making the request.</param>
+        /// <param name="currentUserInfo">The UserInfo object for the current user, to check permissions.</param>
         /// <returns>List of Measurements.</returns>
-        public async Task<List<Measurement>> GetMeasurementsList(int progenyId, int accessLevel)
+        public async Task<List<Measurement>> GetMeasurementsList(int progenyId, UserInfo currentUserInfo)
         {
-            List<Measurement> measurementsList = await GetMeasurementsListFromCache(progenyId);
-            if (measurementsList.Count == 0)
+            MeasurementsListCacheEntry cacheEntry = await _kinaUnaCacheService.GetMeasurementsListCache(currentUserInfo.UserId, progenyId);
+            TimelineUpdatedCacheEntry timelineUpdatedCacheEntry = await _kinaUnaCacheService.GetProgenyOrFamilyTimelineUpdatedCache(progenyId, 0, KinaUnaTypes.TimeLineType.Measurement);
+            if (cacheEntry != null && timelineUpdatedCacheEntry != null)
+            {
+                if (cacheEntry.UpdateTime >= timelineUpdatedCacheEntry.UpdateTime)
+                {
+                    return cacheEntry.MeasurementsList.ToList();
+                }
+            }
+
+            Measurement[] measurementsList = await GetMeasurementsListFromCache(progenyId);
+            if (measurementsList.Length == 0)
             {
                 measurementsList = await SetMeasurementsListInCache(progenyId);
             }
 
-            measurementsList = [.. measurementsList.Where(m => m.AccessLevel >= accessLevel)];
+            List<Measurement> accessibleMeasurements = [];
+            foreach (Measurement measurement in measurementsList)
+            {
+                if (await _accessManagementService.HasItemPermission(KinaUnaTypes.TimeLineType.Measurement, measurement.MeasurementId, currentUserInfo, PermissionLevel.View))
+                {
+                    measurement.ItemPerMission = await _accessManagementService.GetItemPermissionForUser(KinaUnaTypes.TimeLineType.Measurement, measurement.MeasurementId, measurement.ProgenyId, 0, currentUserInfo);
+                    accessibleMeasurements.Add(measurement);
+                }
+            }
 
-            return measurementsList;
+            await _kinaUnaCacheService.SetMeasurementsListCache(currentUserInfo.UserId, progenyId, accessibleMeasurements.ToArray());
+            
+            return accessibleMeasurements;
         }
 
         /// <summary>
@@ -172,13 +244,13 @@ namespace KinaUnaProgenyApi.Services
         /// </summary>
         /// <param name="progenyId">The ProgenyId of the Progeny to get Measurements for.</param>
         /// <returns>List of Measurements.</returns>
-        private async Task<List<Measurement>> GetMeasurementsListFromCache(int progenyId)
+        private async Task<Measurement[]> GetMeasurementsListFromCache(int progenyId)
         {
-            List<Measurement> measurementsList = [];
+            Measurement[] measurementsList = [];
             string cachedMeasurementsList = await _cache.GetStringAsync(Constants.AppName + Constants.ApiVersion + "measurementslist" + progenyId);
             if (!string.IsNullOrEmpty(cachedMeasurementsList))
             {
-                measurementsList = JsonConvert.DeserializeObject<List<Measurement>>(cachedMeasurementsList);
+                measurementsList = JsonSerializer.Deserialize<Measurement[]>(cachedMeasurementsList, JsonSerializerOptions.Web);
             }
 
             return measurementsList;
@@ -189,10 +261,10 @@ namespace KinaUnaProgenyApi.Services
         /// </summary>
         /// <param name="progenyId">The ProgenyId of the Progeny to get Measurements for.</param>
         /// <returns>List of Measurements.</returns>
-        private async Task<List<Measurement>> SetMeasurementsListInCache(int progenyId)
+        private async Task<Measurement[]> SetMeasurementsListInCache(int progenyId)
         {
-            List<Measurement> measurementsList = await _context.MeasurementsDb.AsNoTracking().Where(m => m.ProgenyId == progenyId).ToListAsync();
-            await _cache.SetStringAsync(Constants.AppName + Constants.ApiVersion + "measurementslist" + progenyId, JsonConvert.SerializeObject(measurementsList), _cacheOptionsSliding);
+            Measurement[] measurementsList = await _context.MeasurementsDb.AsNoTracking().Where(m => m.ProgenyId == progenyId).ToArrayAsync();
+            await _cache.SetStringAsync(Constants.AppName + Constants.ApiVersion + "measurementslist" + progenyId, JsonSerializer.Serialize(measurementsList, JsonSerializerOptions.Web), _cacheOptionsSliding);
             return measurementsList;
         }
     }
